adding some syscall abi for x86_64 , and fixing unhandled exception on boot:x64 mode

+++ os/kernel/linux_compat/linux_syscall_x86_64.h
@@ -5,9 +5,9 @@
 extern "C" {
 #endif
 
-int linux_syscall_dispatch_x86_64(uint32_t num, uint32_t a1, uint32_t a2,
-                                  uint32_t a3, uint32_t a4, uint32_t a5,
-                                  uint32_t a6);
+int linux_syscall_dispatch_x86_64(uint64_t num, uint64_t a1, uint64_t a2,
+                                  uint64_t a3, uint64_t a4, uint64_t a5,
+                                  uint64_t a6);
 
 #ifdef __cplusplus
 }
